requirements.txt
torch>=1.9.0
transformers>=4.20.0
pillow>=8.0.0
open_clip_torch>=2.26.1
hydra-core
omegaconf
wandb
pandas
tqdm
openai
anthropic
tenacity
huggingface_hub
accelerate
einops
transformers_stream_generator
torchvision
tiktoken
matplotlib
timeout-decorator